Move common inside asso
