+++ accountant/models.py
@@ -11,8 +11,8 @@ from django.utils.translation import gettext_lazy as _
 from djmoney.models.fields import MoneyField
 
 # Project Import
-from common import models as cm
-from common import utils as u
+from asso.common import models as cm
+from asso.common import utils as u
 
 
 class Account(cm.Base, cm.EditInfo, cm.TrashBin):
