+++ asso/common/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class CommonConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "common"
+    name = "asso.common"
