+++ asso/common/views.py
